fix(seo): page-number → real 308; unknown-slug → 404 (not redirect to /)
Two leftovers from PR #2002:

1) /book/{slug}/page-number/{n} now returns HTTP 308 to /page/{pageId}.
   The page.tsx version called permanentRedirect(), but under Next 16
   RSC/streaming that emits 200 + a NEXT_REDIRECT marker in the body.
   Browsers handle it; crawlers don't, so Google logged these as
   "Soft 404" / "Page with redirect" despite the intended 307. Replacing
   page.tsx with route.ts (NextResponse.redirect, status 308) gives us a
   real HTTP-level redirect that transfers signal to the canonical URL.

2) src/proxy.ts: unknown root slugs now 404 instead of 307 → /. Spam
   scrapers link bogus URLs (/tag/phpmyadmin/, /walking-with-confidence/,
   …) at the domain by the thousand; the redirect-to-home pattern showed
   up in GSC as "Soft 404" / "Page with redirect" and Google kept
   recrawling them. 404 with X-Robots-Tag: noindex lets Google drop the
   entries permanently on next visit. Matches user expectation too: a
   typo'd URL should say "not found", not silently jump to the homepage.
